separate positions file and max delta
